Bug fixing major and minon
